% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-latte.R
\name{call_latte}
\alias{call_latte}
\title{Unified interface to calling LattE}
\usage{
call_latte(
  which_exe,
  input_file,
  project_dir,
  opts = list(),
  stdout = "stdout",
  stderr = "stderr",
  shell = FALSE
)
}
\arguments{
\item{which_exe}{The name of LattE executable, e.g. \code{"count"}}

\item{input_file}{The filename of file to be given to the executable, without
a path}

\item{project_dir}{The path associated with \code{input_file}.}

\item{opts}{A named list of optional command line arguments to the
executable.}

\item{stdout, stderr}{Filenames to redirect standard output and standard error
to.}

\item{shell}{\code{logical(1)}; should the shell command be messaged to the
user?}
}
\value{
\code{invisible()}
}
\description{
Unified interface to calling LattE
}
