% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latte.R
\docType{package}
\name{latte}
\alias{latte}
\alias{package-latte}
\alias{latte-package}
\title{R Interface to LattE and 4ti2}
\description{
Back-end connections to LattE (\url{https://www.math.ucdavis.edu/~latte/})
and 4ti2 (\url{http://www.4ti2.de/}) executables and front-end tools
facilitating their use in the R ecosystem.
}
