% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latte_files.R
\name{latteFiles}
\alias{latteFiles}
\alias{format_latte}
\alias{write_latte}
\alias{write.latte}
\alias{read_latte}
\alias{read.latte}
\title{Format/read/write a matrix in latte's style}
\usage{
format_latte(mat, file)

write_latte(mat, file)

write.latte(mat, file)

read_latte(file, format = c("mat", "Ab"))

read.latte(file, format = c("mat", "Ab"))
}
\arguments{
\item{mat}{A matrix}

\item{file}{A filename}

\item{format}{"mat" or "Ab"}
}
\value{
\itemize{

\item \code{\link[=format_latte]{format_latte()}} -- A character string of the matrix in
latte format.

\item \code{\link[=write.latte]{write.latte()}} -- An invisible character
string of the formatted output.

\item \code{\link[=read.latte]{read.latte()}} -- An integer matrix.

}
}
\description{
\code{format_latte} formats a matrix in latte's style.
\code{write.latte} writes a latte-formatted file to file.
\code{read.latte} reads a latte-formatted file from disk.
}
\examples{

\dontrun{ creates files

(mat <- matrix(sample(9), 3, 3))

format_latte(mat)
cat(format_latte(mat))

write.latte(mat, "foo.hrep")
file.show("foo.hrep")
read.latte("foo.hrep")
read.latte("foo.hrep", "Ab")

attr(mat, "linearity") <- c(1, 3)
attr(mat, "nonnegative") <- 2
mat
format_latte(mat)
cat(format_latte(mat))
write.latte(mat, "foo.hrep")
file.show("foo.hrep")
read.latte("foo.hrep")

file.remove("foo.hrep")

}


}
\seealso{
\code{\link[=write.latte]{write.latte()}}
}
