% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latter.R
\docType{package}
\name{latter}
\alias{latter}
\alias{package-latter}
\alias{latter-package}
\title{LattE and 4ti2 in R}
\description{
Back-end connections to LattE (\url{https://www.math.ucdavis.edu/~latte/})
and 4ti2 (\url{http://www.4ti2.de/}) executables and front-end tools
facilitating its use in the R ecosystem.
}
