% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ones.R
\name{ones}
\alias{ones}
\alias{ones_r}
\title{Make a vector of ones}
\usage{
ones(n)

ones_r(n)
}
\arguments{
\item{n}{How many ones}
}
\value{
A Column/row vector of ones as an integer matrix
}
\description{
\code{ones} makes a column vector of ones, \code{ones_r} makes a
row vector of one.
}
\examples{

ones(5)
ones_r(5)

str(ones(5))

}

