% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matrix.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot a matrix}
\usage{
plot_matrix(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
a ggplot object
}
\description{
plot_matrix is a R variant of Matlab's spy function.
}
\examples{

# the no-three-way interaction configuration
A <- hmat(c(3,3,3), subsets(1:3, 2))
Reduce(kronecker, list())
kronecker(diag(3), )
kronecker()
plot_matrix(A)
plot_matrix(markov(A))

}
\author{
David Kahle \email{david.kahle@gmail.com}
}

