% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_paths.R
\name{setPaths}
\alias{setPaths}
\alias{set_4ti2_path}
\alias{set_latte_path}
\title{Set paths to external functions}
\usage{
set_latte_path(path)

set_4ti2_path(path)
}
\arguments{
\item{path}{A character string, the path to a 4ti2 function (e.g.
markov) for setting 4ti2's path or a LattE function (e.g.
count) for LattE's path}
}
\value{
An invisible character string, the path found.  More
  importantly, the function has the side effect of setting the
  option "latte_path" or "4ti2_path"
}
\description{
These functions set the path to external programs either by (1)
passing them a character string or (2) using
\code{\link{file.choose}}.
}
\examples{

\dontrun{ # the below code requires suggested external software

## for LattE
getOption("latte_path")
set_latte_path()

## for 4ti2 (typically the same as LattE)
getOption("4ti2_path")
set_4ti2_path()



## each of these functions can be used statically as well
(`4ti2_path` <- getOption("4ti2_path"))
set_4ti2_path("/path/to/4ti2/directory")
getOption("4ti2_path")
set_4ti2_path(4ti2_path) # undoes example



}

}
\author{
David Kahle \email{david.kahle@gmail.com}
}

