% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableau.R
\name{tableau}
\alias{tableau}
\title{Tableau Notation for Markov}
\usage{
tableau(move, dim)
}
\arguments{
\item{move}{a markov move matrix, where the columns are moves in
vector form (e.g. the output of markov)}

\item{dim}{the dimensions of the table form of the move,
oftentimes a vector of the number of levels of each variable in
order}
}
\value{
an object of class tableau
}
\description{
Print the tableau notation for a Markov move.  See the reference
provided, p. 13.
}
\examples{

vec <- matrix(c(1, -1, -1, 1), nrow = 4)
varlvls <- c(2, 2)
tableau(vec, varlvls)


}
\references{
Drton, M., B. Sturmfels, and S. Sullivant (2009).
\emph{Lectures on Algebraic Statistics}, Basel: Birkhauser
Verlag AG.
}
